+++ app/components/chat/KakaoChat.tsx
@@ -10,10 +10,11 @@ declare global {
 
 const KakaoChat = () => {
   useEffect(() => {
-    // Kakao SDK 로드 후에만 실행
+    // Kakao SDK 로드 후에만 실행"
     const initializeKakao = () => {
       if (window.Kakao && !window.Kakao.isInitialized()) {
-        window.Kakao.init(process.env.NEXT_PUBLIC_KAKAO_API_KEY); // 사용하려는 앱의 JavaScript 키 입력
+       // window.Kakao.init(process.env.NEXT_PUBLIC_KAKAO_API_KEY); // 사용하려는 앱의 JavaScript 키 입력
+        window.Kakao.init('wqrewdhfjkdsfhksjgs'); // 사용하려는 앱의 JavaScript 키 입력
       }
 
       // DOM이 렌더링된 후에 채널 버튼 생성
