+++ app/components/common/Row/GroupRow.tsx
@@ -26,6 +26,7 @@ const GroupRow = ({ active, onSelect }: GroupRowProps) => {
 
   useEffect(() => {
     groupService.findList(page, pageSize, dispatch);
+    
   }, [page, pageSize]);
 
   if (loading) return <LoadingSpinner />;
