+++ app/model/user/user.model.ts
@@ -9,7 +9,6 @@ export interface UserModel {
   state?:boolean
   logoutAt?: Date | string;
 }
-
 // 상태 인터페이스 정의
 interface UserState {
   currentUser: UserModel | null;
