--- app/queryClient.tsx
@@ -1,4 +1,0 @@
-// src/queryClient.ts
-import { QueryClient } from 'react-query';
-
-export const queryClient = new QueryClient();
