+++ app/service/group/book.service.ts
@@ -1,13 +1,13 @@
 import { AppDispatch } from "@/lib/store";
 import { saveBooks, saveError, saveLoading, saveTotalPage } from "@/lib/features/group/book.slice";
 import bookAPI from "@/app/api/generate/book.api";
-import { fileService } from "../File/file.service";
+import { fileService } from "../file/file.service";
 import { FileType } from "@/app/model/file/file.model";
 
 
-  
 
-  const findList = async (page: number, size: number, dispatch: AppDispatch): Promise<void> => {
+
+const findList = async (page: number, size: number, dispatch: AppDispatch): Promise<void> => {
     try {
         dispatch(saveLoading(true));
         const response = await bookAPI.findList(page, size);
